chatbot_with_images/requirements.txt
torch==2.6.0
accelerate>=0.26.0
numpy
timm
huggingface_hub
json_repair
matplotlib
einops
seaborn
tiktoken
langchain_community
langchain_google_genai
python-docx
transformers
langgraph
langchain_core
ipykernel
python-dotenv
markdown
requests
google-api-python-client
nest_asyncio
jupyter
youtube_transcript_api
beautifulsoup4
yt-dlp
joblib
opencv-python-headless
faiss-cpu
streamlit
groq